+++ src/elements/schema-form.js
@@ -1,0 +1,10 @@
+import {bindable, bindingMode} from 'aurelia-framework';
+
+export class SchemaFormCustomElement {
+  @bindable schema
+
+  @bindable({defaultBindingMode: bindingMode.twoWay})
+  model
+
+  @bindable messages
+}