+++ src/frameworks/actions.js
@@ -1,0 +1,8 @@
+export class Actions {
+
+  activate(model) {
+    this.attribute = model.attribute;
+    this.model     = model.value;
+  }
+
+}