+++ src/frameworks/bootstrap/form-group.js
@@ -1,0 +1,7 @@
+import {FormGroup} from './../form-group';
+
+/***
+ * in case behaviour has to be altered
+ */
+
+export class FormGroupCustomElement extends FormGroup { }