+++ src/frameworks/collection.js
@@ -1,0 +1,8 @@
+export class Collection {
+
+  activate(model) {
+    this.attribute = model.attribute;
+    this.models    = model.value;
+  }
+
+}