+++ doc/CHANGELOG.md
@@ -1,2 +1,33 @@
+### 0.1.2 (2016-06-06)
+
+
+#### Bug Fixes
+
+* **bootstrap:** button group for actions element ([d517d283](https://github.com/SpoonX/aurelia-form/commit/d517d283cd5492bc5bf74674c4b7bfa6badbf1a6))
+* **dependencies:** i18n as a hard dependency ([9858b963](https://github.com/SpoonX/aurelia-form/commit/9858b963520102b5303754af8e65ac47c401d815))
+* **field:** do not bind model to custom element ([b2609823](https://github.com/SpoonX/aurelia-form/commit/b2609823ec9c9cbec4970a7001b98ab00b45a278))
+* **form:** typo fix ([69abddbd](https://github.com/SpoonX/aurelia-form/commit/69abddbd9aab78c9dfce1178ec14a024524e0bb4))
+* **i18n:** use attribute to enable soft dependency on i18n ([cd9a7098](https://github.com/SpoonX/aurelia-form/commit/cd9a7098b59316e9e1e181fc7c549fc8b7ab9c2c))
+* **labels:** move the labels to the framework components ([9bf0deef](https://github.com/SpoonX/aurelia-form/commit/9bf0deef239a5b6ee3c62606ca3f49f0d7695f9c))
+* **naming:** schema to attribute and name to key ([b2987c9d](https://github.com/SpoonX/aurelia-form/commit/b2987c9d0b0692eed29831377340a137d58bed0f))
+* **view:** name of aurelia-view changed to aurelia-view-manager ([eeb9697c](https://github.com/SpoonX/aurelia-form/commit/eeb9697cb8313e6a7c677d95bd9adae2bbba48cf))
+
+
+#### Features
+
+* **actions:** define action buttons in schema ([f205d9c3](https://github.com/SpoonX/aurelia-form/commit/f205d9c3cf321e01568a5335e42fa6ef25b91a08))
+* **aliases:**
+  * added alias nested = fieldset ([e72c4d3b](https://github.com/SpoonX/aurelia-form/commit/e72c4d3b6b2d5f5aba3bc96eb3f928be6377b1a1))
+  * configure aliases to register aliases for a type ([8ee1dda7](https://github.com/SpoonX/aurelia-form/commit/8ee1dda7ed7fffe2b179e7ca1eb520eb6a5f49be))
+* **bootstrap:** the collection custom component ([9ac7f579](https://github.com/SpoonX/aurelia-form/commit/9ac7f5795abcaa2fcef88fee478eb98c196d2342))
+* **component:**
+  * show warning when component not defined ([272d93c0](https://github.com/SpoonX/aurelia-form/commit/272d93c0f1d21f079944b5dab327491dc3c04064))
+  * default the type to "text" when not a valid type ([81274051](https://github.com/SpoonX/aurelia-form/commit/812740511ebb4960f523231ccbb8bb32a6c2e2c9))
+* **fields:** show warning when model is undefined ([714c3722](https://github.com/SpoonX/aurelia-form/commit/714c3722cc04969fec96abda256bc54b281c2070))
+* **fieldset:** nested schemas fieldset element enables the rendering of nested models/entities ([5d7733a2](https://github.com/SpoonX/aurelia-form/commit/5d7733a24f939e72981a64bff4bf57cce8517b3d))
+* **orm:** function for generating schema from entity ([dd81d9b3](https://github.com/SpoonX/aurelia-form/commit/dd81d9b35dae07fd12989501a1f6e469260b80a2))
+* **view:** let viewManager determine locations of templates ([185e98aa](https://github.com/SpoonX/aurelia-form/commit/185e98aaebfc707d520bcc5a1ee634ce1adc419a))
+
+
 ### 0.0.2 (2015-12-07)
 * Updated to work with new orm usage